Reorganizing repository. Adding versioning and testing
