+++ README.md
@@ -1,8 +1,10 @@
-A python package for creating and manipulating pairtrees
+# pypairtree
 
-https://confluence.ucop.edu/display/Curation/PairTree
+v0.0.1
 
-Early beta - no docs
+[![Build Status](https://travis-ci.org/bnbalsamo/pypairtree.svg?branch=master)](https://travis-ci.org/bnbalsamo/pypairtree) [![Coverage Status](https://coveralls.io/repos/github/bnbalsamo/pypairtree/badge.svg?branch=master)](https://coveralls.io/github/bnbalsamo/pypairtree?branch=master)
 
-Author: Brian Balsamo
-Email: brian [at] brianbalsamo [dot] com
+A python package for creating and manipulating pairtrees.
+
+# Author
+Brian Balsamo <brian@brianbalsamo.com>
