+++ quick_test.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+echo "==> Running Tests <=="
+coverage run -m py.test \
+    && echo "==> Coverage <==" && \
+    coverage report
+echo "==> Flake8 <=="
+flake8
+echo "==> Done <=="
