+++ setup.cfg
@@ -1,0 +1,37 @@
+# Coverage config
+[coverage:run]
+source = pypairtree
+
+# autopep8 config
+[pep8]
+in-place = True
+recursive = True
+max-line-length = 120
+verbose = 1
+exclude = .env, build, dist, .git, __pycache__
+
+# flake8 config
+[flake8]
+max-line-length = 120
+ignore = E121,E123,E126,E226,E24,E402,E704,W503,W504
+exclude = .env, build, dist, .git, __pycache__
+
+# bumpversion config
+[bumpversion]
+current_version = 0.0.1
+commit = True
+tag = True
+
+[bumpversion:file:setup.py]
+
+[bumpversion:file:pypairtree/__init__.py]
+
+[bumpversion:file:README.md]
+
+# bdist_wheel config
+# Uncomment for a universal wheel
+#[bdist_wheel]
+#universal = 1
+
+[metadata]
+license_file = LICENSE
