+++ tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Unit tests for pypairtree
+"""
